+++ src/lib.rs
@@ -37,4 +37,3 @@ pub(crate) mod internal;
 pub mod buf;
 pub mod core;
 pub mod grid;
-pub mod impls;