+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `GridRead` has default implementations of `{row|col|rect}_iter`.
+- `GridWrite` has default implementations of `set_{row|col|rect}`.
+
+### Removed
+
+- Removed unused `impls` module.
+
 ## [0.1.0] - 2025-07-14
 
 ### Added